requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.30
faiss-cpu
streamlit
python-dotenv
tavily-python
pytest
